Rename project to nnbench
Reasons are that mlbench is already taken both on PyPI and GitHub Pages,
ironically by different projects respectively.

nnbench is both short and still conveys its main purpose well enough.
