+++ docs/guides/index.md
@@ -1,4 +1,4 @@
 # User Guide
 
-The mlbench user guide provides documentation for users of the library looking to solve specific tasks.
+The nnbench user guide provides documentation for users of the library looking to solve specific tasks.
 See the [Quickstart guide](../quickstart.md) for an introductory tutorial.
