+++ pyproject.toml
@@ -3,7 +3,7 @@ requires = ["setuptools>=45", "setuptools-scm[toml]>=7.1", "wheel"]
 build-backend = "setuptools.build_meta"
 
 [project]
-name = "mlbench"
+name = "nnbench"
 description = "A small framework for benchmarking machine learning models."
 keywords = ["Benchmarking", "Machine Learning"]
 requires-python = ">=3.9"
@@ -36,10 +36,10 @@ dependencies = []
 dynamic = ["version"]
 
 [project.urls]
-Homepage = "https://github.com/aai-institute/mlbench"
-Repository = "https://github.com/aai-institute/mlbench.git"
-Issues = "https://github.com/aai-institute/mlbench/issues"
-Discussions = "https://github.com/aai-institute/mlbench/discussions"
+Homepage = "https://github.com/aai-institute/nnbench"
+Repository = "https://github.com/aai-institute/nnbench.git"
+Issues = "https://github.com/aai-institute/nnbench/issues"
+Discussions = "https://github.com/aai-institute/nnbench/discussions"
 
 [project.optional-dependencies]
 dev = [
@@ -68,7 +68,7 @@ package-dir = { "" = "src" }
 where = ["src"]
 
 [tool.setuptools.package-data]
-mlbench = ["py.typed"]
+nnbench = ["py.typed"]
 
 # Automatically determine version number from Git tags
 [tool.setuptools_scm]
